requirements.txt
fastapi>=0.111.0
uvicorn
sqlalchemy
asyncpg
alembic
aiogram>=3.0.0
pydantic
python-dotenv
redis
boto3
Babel
uvicorn
httpx
